Update README.md
A dating script is a pre-written software code that serves as the foundation for creating a dating platform. It encompasses all the necessary features and functionalities required for users to interact, match, and communicate with potential partners. These scripts can be customized to cater to specific requirements, allowing developers to build a unique dating app or website without starting from scratch.